RDKCOM-2635: REFPLTB-1398 : Harvester - Resolve all build warnings
Reason for change: Have to clear all warnings with respect to 64bit architecture
Test Procedure: build should be success on both 64bit and 32bit arch
Risks: Low

(cherry picked from commit a56c1d3d91f79687eeb9d024c57f41927e00e5ca)

+++ source/HarvesterSsp/webpa_interface.c
@@ -145,7 +145,7 @@ void sendWebpaMsg(char *serviceName, char *dest, char *trans_id, char *contentTy
         CcspHarvesterConsoleTrace(("RDK_LOG_DEBUG, wrp_msg->msg_type :%d\n",wrp_msg->msg_type));
         if(wrp_msg->u.event.payload!=NULL)
         	CcspHarvesterConsoleTrace(("RDK_LOG_DEBUG, wrp_msg->u.event.payload :%s\n",(char *)(wrp_msg->u.event.payload)));
-        CcspHarvesterConsoleTrace(("RDK_LOG_DEBUG, wrp_msg->u.event.payload_size :%d\n",wrp_msg->u.event.payload_size));
+        CcspHarvesterConsoleTrace(("RDK_LOG_DEBUG, wrp_msg->u.event.payload_size :%lu\n",(ULONG)wrp_msg->u.event.payload_size));
 		if(wrp_msg->u.event.source!=NULL)
         	CcspHarvesterConsoleTrace(("RDK_LOG_DEBUG, wrp_msg->u.event.source :%s\n",wrp_msg->u.event.source));
 		if(wrp_msg->u.event.dest!=NULL)
